Moved #listFilesRecursively to FileExt class
