+++ shared/src/main/java/org/odk/collect/shared/files/DirectoryUtils.kt
@@ -5,18 +5,6 @@ import java.io.File
 object DirectoryUtils {
 
     @JvmStatic
-    fun listFilesRecursively(directory: File): List<File> {
-        val listFiles = directory.listFiles() ?: emptyArray()
-        return listFiles.flatMap {
-            if (it.isDirectory) {
-                listFilesRecursively(it)
-            } else {
-                listOf(it)
-            }
-        }
-    }
-
-    @JvmStatic
     fun deleteDirectory(directory: File) {
         directory.deleteRecursively()
     }
